requirements.txt
torch
torchvision
timm
albumentations
opencv-python
mtcnn
pillow
pandas
scikit-learn
tqdm
